routing set-up and working, have not moved modal to user page
