+++ client/components/user.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class User extends React.Component {
+  render() {
+    return (
+      <div>
+        <div>hello from user page</div>
+      </div>
+    );
+  }
+}
+
+export default User;
