+++ client/index.jsx
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 import App from './components/app';
 
 ReactDOM.render(
-  <App/>,
+  <Router>
+    <App />
+  </Router>,
   document.querySelector('#root')
 );
