+++ package.json
@@ -26,7 +26,9 @@
     "@fortawesome/react-fontawesome": "0.1.8",
     "json-server": "0.15.1",
     "react": "16.12.0",
-    "react-dom": "16.12.0"
+    "react-dom": "16.12.0",
+    "react-router": "5.1.2",
+    "react-router-dom": "5.1.2"
   },
   "devDependencies": {
     "@babel/core": "7.7.5",
